refactor: rename SInputSelectSearch -> SInputAsyncDropdown; Option.ts -> InputDropdown.ts
- `SInputSelectSearch.vue` -> `SInputAsyncDropdown.vue` (the name says what it is: a
  dropdown whose options load asynchronously). Updates LensInputSelect, the spec,
  and the component's own class/CSS + doc references.
- `support/Option.ts` -> `support/InputDropdown.ts` (sits alongside InputBase /
  InputText / InputFileUpload; the exported type names — Option, OptionText, … —
  are unchanged). All importers repointed; the module comment trimmed to match the
  other support/Input* files.
