+++ lib/blocks/lens/components/LensFormFilterCondition.vue
@@ -6,7 +6,7 @@ import SInputDropdown from '../../../components/SInputDropdown.vue'
 import SSpinner from '../../../components/SSpinner.vue'
 import { useTrans } from '../../../composables/Lang'
 import { useValidation } from '../../../composables/Validation'
-import { type Option } from '../../../support/Option'
+import { type Option } from '../../../support/InputDropdown'
 import { required } from '../../../validation/rules'
 import { type FieldData } from '../FieldData'
 import { type FilterOperator, FilterOperatorLabelDict } from '../FilterOperator'
