+++ lib/blocks/lens/components/LensFormFilterGroup.vue
@@ -4,7 +4,7 @@ import IconTrash from '~icons/ph/trash'
 import SButton from '../../../components/SButton.vue'
 import SInputSegments from '../../../components/SInputSegments.vue'
 import { useTrans } from '../../../composables/Lang'
-import { type Option } from '../../../support/Option'
+import { type Option } from '../../../support/InputDropdown'
 import { type FieldData } from '../FieldData'
 import LensFormFilterCondition, { type FilterCondition } from './LensFormFilterCondition.vue'
